requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
distro==1.9.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
mysqlclient==2.2.6
openai==1.59.7
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
